requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
openai==2.30.0
python-dotenv==1.2.2
pydantic==2.12.5
mysql-connector-python==9.2.0
